Start the extraction of the presenter code to a gem called Showtime
