+++ lib/showtime.rb
@@ -1,0 +1,2 @@
+require_relative "showtime/presenter"
+require_relative "showtime/view_context"
