+++ lib/showtime/view_context.rb
@@ -1,4 +1,4 @@
-module Presenter
+module Showtime
   module ViewContext
     def view_context
       properties[:view_context]
